Added 'unit_tests' and 'integration_tests' directories inside '__test__' directory and sorted all previous tests appropiately
