--- src/__test__/pages/chapters/chaptersContainer.test.js
@@ -1,22 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import ChaptersContainer from '../../../js/pages/chapters/ChaptersContainer';
-import configureMockStore from 'redux-mock-store';
-import thunk from 'redux-thunk';
-
-const middlewares = [thunk]
-const mockStore = configureMockStore(middlewares)
-
-describe('ChaptersContainer', () => {
-    const store = mockStore({
-        direction:{},
-        geolocation: {},
-        chaptersContainer: {
-        }
-    });
-    const wrapper = shallow(<ChaptersContainer store={store} />);
-    it('should render', () => {
-        expect(wrapper).toHaveLength(1);
-    });
-
-});
