--- src/__test__/pages/takes/components/TakeList.test.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import TakeList from '../../../../js/pages/takes/TakeList';
-
-describes('Test set rating function', () => {
-
-
-    //const wrapper = shallow(<TakeList />);
-
-   // const test = wrapper.instance().onDrop();
-
-    ///console.log(test);
-
-  //  expect(true).toEqual(true);
-
-});
